REFACTOR :: features, api 대통합
