+++ src/server/notice/queries.ts
@@ -2,6 +2,7 @@ import { noticeList } from "@/api/notice";
 import * as KEY from "@/constants/key";
 import { useQuery } from "react-query";
 
+// 공지사항 리스트
 interface NoticeListType {
   id: number;
   title: string;
